src/requirements.txt
requests
python-dotenv
openai>=1.59.7
pandas
azure-cosmos
semantic-kernel>=1.37.0
azure-ai-projects>=1.1.0b4
azure-ai-agents 
azure-identity 
pyodbc
python-dotenv
azure-kusto-data
opentelemetry-sdk
azure-core-tracing-opentelemetry
azure-monitor-opentelemetry
opentelemetry-instrumentation-openai-v2
Pillow
sqlalchemy
opencv-python
azure-search-documents
fastapi
uvicorn[standard]
azure.ai.inference
orjson
numpy
azure-ai-evaluation[redteam]>=1.11.1
a2a-sdk[sqlite]>=0.2.9
sse-starlette>=2.2.1
starlette>=0.46.1